1. Modify to pass USB-IF CV - Chapter 9 & Class test. 2. SPI driver: Add SPI_TRIGGER_TX_RX_PDMA and SPI_DISABLE_TX_RX_PDMA API. 3. Correct USBD_VCOM_DualPort Descriptor.

+++ SampleCode/StdDriver/SPI_SlaveMode/main.c
@@ -148,7 +148,7 @@ void SPI_Init(void)
     /*---------------------------------------------------------------------------------------------------------*/
     /* Configure SPI0 as a slave, clock idle low, 32-bit transaction, drive output on falling clock edge and latch input on rising edge. */
     /* Configure SPI0 as a low level active device. SPI peripheral clock rate = f_PCLK0 */
-    SPI_Open(SPI0, SPI_SLAVE, SPI_MODE_0, 32, NULL);
+    SPI_Open(SPI0, SPI_SLAVE, SPI_MODE_0, 32, (uint32_t)NULL);
 }
 
 /*** (C) COPYRIGHT 2014~2015 Nuvoton Technology Corp. ***/
